Removes experimental worker managed asynchronous approach
Previously, an experimental approach to handle asynchronous operations
for iOS was implemented that leveraged Kotlin workers and required
all calls to be made from the main thread. This approach produced
valid, asynchronous operations but muddied the waters given the
nature of this library. The simple approach, using the iOS main thread
for coroutines, is preferred here and a recommendation is provided to
use synchronous methods until multithreaded coroutines are fully
supported in Kotlin/Native.